Initial commit on android_pytools
A few of the commonly used adb commands have been wrapped. Any
exception that is thrown during the running of a wrapped ADB
command is passed to the calling function as a ADBCommandException. The
results of an ADB command are encapsulated in a ADBCommandResult  object and
returned to the calling function.