Renamed and refactored ADB exceptions into an exceptions module.
ADBExecutionError is now called ADBProcessError to align with the intent
of it being thrown when a CalledProcessError occurs. This exception was
moved to exceptions.py where it should live with other ADB specific
exceptions.

ADBException was created as a generic exception to throw for any ADB
related issues. ADBExecutionError now inherits from ADBException since
it is a specialized version of ADBException.